firmware: uint32_t -> uintpr_t

+++ flight/PiOS/Boards/STM32103CB_PIPXTREME_Rev1.h
@@ -184,7 +184,7 @@ extern uintptr_t pios_com_telem_uart_flexi_id;
 extern uintptr_t pios_com_telemetry_id;
 extern uintptr_t pios_com_rfm22b_id;
 extern uintptr_t pios_com_radio_id;
-extern uint32_t pios_ppm_rcvr_id;
+extern uintptr_t pios_ppm_rcvr_id;
 #define PIOS_COM_TELEM_USB         (pios_com_telem_usb_id)
 #define PIOS_COM_TELEM_VCP         (pios_com_telem_vcp_id)
 #define PIOS_COM_TELEM_UART_FLEXI  (pios_com_telem_uart_flexi_id)
