+++ flight/targets/board_hw_defs/revolution/board_hw_defs.c
@@ -685,7 +685,7 @@ static const struct pios_overo_cfg pios_overo_cfg = {
 		},
 	} },
 };
-uint32_t pios_overo_id = 0;
+uintptr_t pios_overo_id = 0;
 void PIOS_OVERO_irq_handler(void)
 {
 	/* Call into the generic code to handle the IRQ for this specific device */
